setting ts to js
