+++ package.json
@@ -5,18 +5,21 @@
   "author": "awesome-cis",
   "license": "MIT",
   "scripts": {
-    "dev": "node server.js",
-    "build": "next build",
-    "start": "cross-env NODE_ENV=production node server.js"
+    "build-ts": "tsc --project tsconfig.server.json",
+    "build": "next build && yarn run build-ts",
+    "start": "node dist/server.js",
+    "start:prod": "cross-env NODE_ENV=production node dist/server.js"
   },
   "devDependencies": {
+    "@types/express": "^4.16.1",
+    "@types/next": "^8.0.5",
+    "@types/node": "^12.0.1",
+    "@types/react": "^16.8.17",
+    "@types/react-dom": "^16.8.4",
     "cross-env": "^5.2.0",
     "next": "^8.1.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "@types/next": "^8.0.5",
-    "@types/react": "^16.8.17",
-    "@types/react-dom": "^16.8.4",
     "typescript": "^3.4.5"
   },
   "dependencies": {
