--- server.ts
@@ -1,20 +1,0 @@
-const express = require("express");
-const next = require("next");
-
-const port = parseInt(process.env.PORT || "3000", 10);
-const dev = process.env.NODE_ENV !== "production";
-const app = next({ dev });
-const handle = app.getRequestHandler();
-
-app.prepare().then(() => {
-  const server = express();
-
-  server.get("*", (req: any, res: any) => {
-    return handle(req, res);
-  });
-
-  server.listen(port, (err: any) => {
-    if (err) throw err;
-    console.log(`> Ready on http://localhost:${port}`);
-  });
-});
